kafka-consumer/requirements.txt
kafka-python==2.0.2
google-cloud-bigquery==3.16.0
python-dotenv==1.0.1
